Refactor replace_arith assists into one module
